Moving more code to ingress. Renaming host.py to host_repository.
