+++ app/queue/ingress.py
@@ -1,7 +1,9 @@
 import json
 from marshmallow import Schema, fields, ValidationError
-from app.logging import get_logger
+from app.logging import get_logger, threadctx
+from app.exceptions import InventoryException, ValidationException
 from app.queue import metrics
+from lib import host_repository
 
 
 logger = get_logger(__name__)
@@ -11,6 +13,7 @@ class OperationSchema(Schema):
     operation = fields.Str()
     request_id = fields.Str()
     data = fields.Dict()
+
 
 @metrics.ingress_message_parsing_time.time()
 def parse_operation_message(message):
@@ -36,3 +39,41 @@ def parse_operation_message(message):
         raise
     
     return parsed_operation
+
+
+def add_host(host_data):
+    try:
+        logger.info("Attempting to add host...")
+        host_repository.add_host(host_data)
+        metrics.add_host_success.inc()
+        logger.info("Host added") # This definitely needs to be more specific (added vs updated?)
+    except InventoryException as e:
+        logger.exception("Error adding host ", extra={"host": host_data})
+        metrics.add_host_failure.inc()
+    except Exception as e:
+        logger.exception("Error while adding host", extra={"host": host_data})
+        metrics.add_host_failure.inc()
+
+
+def handle_message(message):
+    validated_operation_msg = parse_operation_message(message)
+    initialize_thread_local_storage(validated_operation_msg)
+    # FIXME: verify operation type
+    add_host(validated_operation_msg["data"])
+
+
+def event_loop(consumer, flask_app, handler=handle_message):
+    with flask_app.app_context():
+        logger.debug("Waiting for message")
+        for msg in consumer:
+            logger.debug("Message received")
+            try:
+                handler(msg)
+                metrics.ingress_message_handler_success.inc()
+            except Exception:
+                metrics.ingress_message_handler_failure.inc()
+                logger.exception("Unable to process message")
+
+
+def initialize_thread_local_storage(operation_message):
+    threadctx.request_id = operation_message.get("request_id", "-1")
