+++ test_host_dedup_logic.py
@@ -2,7 +2,7 @@ import uuid
 
 from app import db
 from app.models import Host
-from lib.host import find_existing_host
+from lib.host_repository import find_existing_host
 from test_utils import flask_app_fixture
 
 ACCOUNT_NUMBER = "000102"
